feat: final documentation and cleanup [RQ-001]
