+++ released_plugins/v3d_plugins/MIND/README.md
@@ -1,0 +1,40 @@
+# Machine Intelligence for Neuron Detection
+
+## Introduction
+
+Machine Intelligence for Neuron Detection (MIND) is a platform built with the aim of facilitating deep learning pipelines to automatically segment neuron somas in large scale images of the brain. Its capabilities and how to use them are clearly documented on its [wiki page](https://github.com/ImagiNeuron/vaa3d_tools/wiki). They include the following functions:
+
+1. `Isotropic Correction` - adjust the 3D view of images for the true dimensions of a voxel based on the microscopy technique.
+2. `Soma Segmentation` - automatically segments neuron somas that have been bounded within spheres around their center of mass with Vaa3D markers. Ouputs a visualization of the segmentation, as well as the binary segmentation file, and a probability model of Neuron Soma sphape obtained through Principal Component Analysis (PCA) on the binary segmentation of somas. The PCA results are also output in a csv file.
+3. `PC Analysis` - Intensity Weighted Principal Component Analysis on the original image based on the spherical volume captured by the Vaa3D markers. Results are output in a csv file.
+4. `Visualize PCA` - Visualize the results of PCA from the `Soma Segmentation` Function in 3D. The longest PC is shown in green, the second longest in red, and the shortest in cyan.
+5. `Visualize Probability Model` - Visualize the probability model output by the Soma Segmenation function in 2D or 3D, as a heatmap with a legend.
+6. `Probability Model Legend` - View only the legend of the probability model.
+7. `Create Background` - Create a synthetic background for somas based on analysing an open image. Gives access to the synthetic background without having to put somas in it.
+8. `Simulate Somas` - Must be run after the `Soma Segmentation` function. Generates sythnetic somas in a sythetic background, along with a corresponding binary segmentatoin. Outputs these as images, along with a csv file containing their PCA results, a marker file locating sythetic somas, an a visualization of the simulation results.
+
+## Architecture
+
+The initial code for this plugin was generated with the `_Vaa3D_plugin_creator` plugin. The following contains a breif description of all the files that contian C++ code, and build files.
+
+### C++ Files
+
+- `soma_segmentation_plugin.cpp` - This file was the first autogenerated by the Vaa3D plugin creator. It controls which functions in the code to call based on the plugin function from above that a user has selected. It also contains the logic for the `Isotropic Correction` function.
+- `ResolutionDialogue.cpp` - This file contains the user input window for the `Isotropic Correction` function.
+- `pcaAnalysis.cpp` - This file contains all the code related to Principal Component Anlysis logic used by the `PC Analysis` and `Soma Segmentation` plugin functions.
+- `cellSegmentation_plugin.h` - This file was adopted from the examplar based cellseg_quickfind plugin, and vastly improved. It contains the logic for the `Soma Segmentation` function, which does segmentation using flood filling and local Otsu thresholding. The files also contains the subsequent logic for doing PCA on the resulting binary segmentation and building a probability model, and saving and visualizing all the associated results.
+- `probability_model_vis.cpp` - This file contains the logic for visualising the probability model output by the `Soma Segmentation` function.
+- `soma_simulation.cpp` - This file contains all the code for the simulation for the `Create Background` and `Simulate Somas` functions.
+
+### Build Files
+
+- All build files in this repository, including the .pro files and the Makefile, were created by the `_Vaa3D_plugin_creator` plugin. We made minor modifications to them to adjust them for the tools we are using, but nothing significant. To build our plugin, simply run the `compile_mind.bat` script under the `\vaa3d_tools\released_plugins` folder.
+
+## Contributions
+
+This project was developed by ImagiNeuron, a group of 4 McGill University students. 
+
+- Shidan Javaheri
+- Siger Ma
+- Athmane Benarous
+- Thibaut Baguette
