+++ released_plugins/v3d_plugins/MIND/object_script.soma_segmentation.Release
@@ -1,6 +1,8 @@
 release/soma_segmentation_plugin.o
 release/ResolutionDialog.o
 release/pcaAnalysis.o
+release/probability_model_vis.o
+release/soma_simulation.o
 release/v3d_message.o
 release/v_neuronswc.o
 release/basic_surf_objs.o
