+++ released_plugins/v3d_plugins/MIND/pcaAnalysis.cpp
@@ -1,7 +1,26 @@
+ /**
+ * 2025-04-18: by ImagiNeuron: Shidan Javaheri, Siger Ma, Athmane Benarous and
+ * Thibaut Baguette (McGill University)
+ */
 #include "pcaAnalysis.h"
 
-// Function to perform PCA analysis
-
+/**
+ * @brief Function to save PCA results to a CSV file
+ *
+ * @param filename The name of the file to save the results to
+ * @param somaIndex The index of the soma being analyzed
+ * @param lm The location of the soma
+ * @param pc1 The first principal component
+ * @param pc2 The second principal component
+ * @param pc3 The third principal component
+ * @param vec1 The first eigenvector
+ * @param vec2 The second eigenvector
+ * @param vec3 The third eigenvector
+ * @param x_center The x-coordinate of the center of mass
+ * @param y_center The y-coordinate of the center of mass
+ * @param z_center The z-coordinate of the center of mass
+ * @param parent The parent widget for the file dialog
+ */
 void savePCAResultsToCSV(const QString &filename, int somaIndex,
                          const LocationSimple &lm, double pc1, double pc2,
                          double pc3, const double *vec1, const double *vec2,
@@ -56,6 +75,26 @@ void savePCAResultsToCSV(const QString &filename, int somaIndex,
          actualFilename.toStdString().c_str());
 }
 
+/**
+ * * @brief Function to analyze a soma using PCA and save results to a CSV file
+ *
+ * @param labeledData The labeled data of the image
+ * @param N The width of the image
+ * @param M The height of the image
+ * @param P The depth of the image
+ * @param lm The location of the soma
+ * @param somaIndex The index of the soma being analyzed
+ * @param savePath The path to save the results
+ * @param pc1 The pointer to the first principal component
+ * @param pc2 The pointer to the second principal component
+ * @param pc3 The pointer to the third principal component
+ * @param vec1 The pointer to the first eigenvector
+ * @param vec2 The pointer to the second eigenvector
+ * @param vec3 The pointer to the third eigenvector
+ * @param x_center The pointer to the x-coordinate of the center of mass
+ * @param y_center The pointer to the y-coordinate of the center of mass
+ * @param z_center The pointer to the z-coordinate of the center of mass
+ */
 void analyzeSomaPCAReturnResults(unsigned char *labeledData, V3DLONG N,
                                  V3DLONG M, V3DLONG P, const LocationSimple &lm,
                                  int somaIndex, QString savePath, double &pc1,
@@ -98,6 +137,12 @@ void analyzeSomaPCAReturnResults(unsigned char *labeledData, V3DLONG N,
   delete[] img3d;
 }
 
+/**
+ * @brief Function to modify file path for TeraFly data structure
+ *
+ * @param fileName The original file name
+ * @return The modified file path
+ */
 QString modifyFilePathForTeraFly(const QString &fileName) {
   QString modifiedFilePath = fileName;
 
@@ -128,6 +173,16 @@ QString modifyFilePathForTeraFly(const QString &fileName) {
   return modifiedFilePath;
 }
 
+/**
+ * * @brief Function to load segmentation file
+ *
+ * @param imageName The name of the image
+ * @param segData The segmentation data
+ * @param sz The size of the image
+ * @param datatype The data type of the image
+ * @param callback The V3D plugin callback interface
+ * @param parent The parent interface
+ */
 void loadSegmentationFile(const QString &imageName, unsigned char *&segData,
                           V3DLONG sz[4], int &datatype,
                           V3DPluginCallback2 &callback, QWidget *parent) {
@@ -159,6 +214,18 @@ void loadSegmentationFile(const QString &imageName, unsigned char *&segData,
   }
 }
 
+/**
+ * * @brief Function to analyze a soma using PCA and save results to a CSV file,
+ * wihtout returning the results
+ *
+ * @param labeledData The labeled data of the image
+ * @param N The width of the image
+ * @param M The height of the image
+ * @param P The depth of the image
+ * @param lm The location of the soma
+ * @param somaIndex The index of the soma being analyzed
+ * @param savePath The path to save the results
+ */
 void analyzeSomaPCA(unsigned char *labeledData, V3DLONG N, V3DLONG M, V3DLONG P,
                     const LocationSimple &lm, int somaIndex, QString savePath) {
   // Extract soma info
@@ -201,6 +268,12 @@ void analyzeSomaPCA(unsigned char *labeledData, V3DLONG N, V3DLONG M, V3DLONG P,
   delete[] img3d;
 }
 
+/**
+ * @brief Function to visuzalize PCA results
+ *
+ * @param callback The V3D plugin callback interface
+ * @param parent The parent interface
+ */
 void visualizePCA_func(V3DPluginCallback2 &callback, QWidget *parent) {
   v3dhandle curwin = callback.currentImageWindow();
   if (!curwin) {
@@ -226,7 +299,8 @@ void visualizePCA_func(V3DPluginCallback2 &callback, QWidget *parent) {
   const int size = p4DImage->getTotalUnitNumber() * p4DImage->getUnitBytes();
   memcpy(pcaVisualization->getRawData(), p4DImage->getRawData(), size);
   memcpy(pcaVisualization->getRawData() + size, p4DImage->getRawData(), size);
-  memcpy(pcaVisualization->getRawData() + 2 * size, p4DImage->getRawData(), size);
+  memcpy(pcaVisualization->getRawData() + 2 * size, p4DImage->getRawData(),
+         size);
 
   // load pca data from csv
   QString filename = QFileDialog::getOpenFileName(parent, "Open PCA Results",
@@ -316,7 +390,8 @@ void visualizePCA_func(V3DPluginCallback2 &callback, QWidget *parent) {
       col++;
     }
 
-    // Visualize: green longest, red second, blue (cyan for better visibility) third
+    // Visualize: green longest, red second, blue (cyan for better visibility)
+    // third
     drawLine(pcaVisualization, 0, 255, 0, center, vec1Pos);
     drawLine(pcaVisualization, 255, 0, 0, center, vec2Pos);
     drawLine(pcaVisualization, 0, 255, 255, center, vec3Pos);
@@ -329,8 +404,18 @@ void visualizePCA_func(V3DPluginCallback2 &callback, QWidget *parent) {
   callback.setImage(newwin, pcaVisualization);
 }
 
-void drawLine(Image4DSimple *image, unsigned char r, unsigned char g, unsigned char b, double *from, double *to)
-{
+/**
+ * @brief Function to draw a line in the image
+ *
+ * @param image - the image to draw on
+ * @param r - red color value
+ * @param g - green color value
+ * @param b - blue color value
+ * @param from - starting point of the line
+ * @param to - ending point of the line
+ */
+void drawLine(Image4DSimple *image, unsigned char r, unsigned char g,
+              unsigned char b, double *from, double *to) {
   // convert points from world space to image space
   int x1 = (int)((from[0] - image->getOriginX()) / image->getRezX());
   int y1 = (int)((from[1] - image->getOriginY()) / image->getRezY());
@@ -342,9 +427,12 @@ void drawLine(Image4DSimple *image, unsigned char r, unsigned char g, unsigned c
   unsigned char *imgData = image->getRawData();
 
   auto fillPixel = [&](int x, int y, int z) {
-    if (x >= 0 && x < image->getXDim() && y >= 0 && y < image->getYDim() && z >= 0 && z < image->getZDim()) {
-      const int channel_size = image->getXDim() * image->getYDim() * image->getZDim();
-      const int subindex = z * image->getXDim() * image->getYDim() + y * image->getXDim() + x;
+    if (x >= 0 && x < image->getXDim() && y >= 0 && y < image->getYDim() &&
+        z >= 0 && z < image->getZDim()) {
+      const int channel_size =
+          image->getXDim() * image->getYDim() * image->getZDim();
+      const int subindex =
+          z * image->getXDim() * image->getYDim() + y * image->getXDim() + x;
       imgData[subindex] = r;
       imgData[channel_size + subindex] = g;
       imgData[2 * channel_size + subindex] = b;
@@ -419,353 +507,4 @@ void drawLine(Image4DSimple *image, unsigned char r, unsigned char g, unsigned c
       fillPixel(x1, y1, z1);
     }
   }
-}
-
-unsigned char *create_background(V3DPluginCallback2 &callback, QWidget *parent,
-                                 V3DLONG &dim_X, V3DLONG &dim_Y,
-                                 V3DLONG &dim_Z) {
-  // Get current image
-  v3dhandle curwin = callback.currentImageWindow();
-  if (!curwin) {
-    v3d_msg("No image opened.", parent);
-    return nullptr;
-  }
-
-  Image4DSimple *p4DImage = callback.getImage(curwin);
-  if (!p4DImage) {
-    v3d_msg("No image opened.", parent);
-    return nullptr;
-  }
-
-  // Get image dimensions and data
-  dim_X = p4DImage->getXDim();
-  dim_Y = p4DImage->getYDim();
-  dim_Z = p4DImage->getZDim();
-  V3DLONG totalSize = dim_X * dim_Y * dim_Z;
-  unsigned char *originalData = p4DImage->getRawData();
-
-  // Calculate chunk dimensions (ensure at least 1)
-  V3DLONG factor = 4;
-  V3DLONG chunk_X = std::max(1L, dim_X / factor);
-  V3DLONG chunk_Y = std::max(1L, dim_Y / factor);
-  V3DLONG chunk_Z = std::max(1L, dim_Z / factor);
-
-  // Calculate number of chunks in each dimension
-  V3DLONG num_chunks_X = (dim_X + chunk_X - 1) / chunk_X;
-  V3DLONG num_chunks_Y = (dim_Y + chunk_Y - 1) / chunk_Y;
-  V3DLONG num_chunks_Z = (dim_Z + chunk_Z - 1) / chunk_Z;
-
-  printf("Image dimensions: %ld x %ld x %ld\n", dim_X, dim_Y, dim_Z);
-  printf("Chunk dimensions: %ld x %ld x %ld\n", chunk_X, chunk_Y, chunk_Z);
-  printf("Number of chunks: %ld x %ld x %ld = %ld chunks total\n", num_chunks_X,
-         num_chunks_Y, num_chunks_Z,
-         num_chunks_X * num_chunks_Y * num_chunks_Z);
-
-  // Allocate 3D array for background
-  unsigned char *backgroundArray = new unsigned char[totalSize];
-
-  // Seed random generator
-  std::srand(std::time(nullptr));
-
-  // Store statistics for each chunk for blending
-  std::vector<std::vector<std::vector<std::pair<double, double>>>> chunkStats(
-      num_chunks_Z, std::vector<std::vector<std::pair<double, double>>>(
-                        num_chunks_Y, std::vector<std::pair<double, double>>(
-                                          num_chunks_X, {0.0, 0.0})));
-
-  // Calculate statistics for each chunk first
-  for (V3DLONG cz = 0; cz < num_chunks_Z; cz++) {
-    V3DLONG z_start = cz * chunk_Z;
-    V3DLONG z_end = std::min(z_start + chunk_Z, dim_Z);
-
-    for (V3DLONG cy = 0; cy < num_chunks_Y; cy++) {
-      V3DLONG y_start = cy * chunk_Y;
-      V3DLONG y_end = std::min(y_start + chunk_Y, dim_Y);
-
-      for (V3DLONG cx = 0; cx < num_chunks_X; cx++) {
-        V3DLONG x_start = cx * chunk_X;
-        V3DLONG x_end = std::min(x_start + chunk_X, dim_X);
-
-        // Create histogram for this chunk
-        int hist[256] = {0};
-        int chunkSize = 0;
-
-        for (V3DLONG z = z_start; z < z_end; z++) {
-          for (V3DLONG y = y_start; y < y_end; y++) {
-            for (V3DLONG x = x_start; x < x_end; x++) {
-              V3DLONG idx = z * dim_X * dim_Y + y * dim_X + x;
-              hist[originalData[idx]]++;
-              chunkSize++;
-            }
-          }
-        }
-
-        // Calculate Otsu threshold for this chunk using the new function
-        int threshold = calculateOtsuThreshold(hist, chunkSize);
-
-        // Calculate statistics for background voxels (below threshold)
-        double mean = 0.0;
-        double stdDev = 0.0;
-        V3DLONG backgroundCount = 0;
-        double sum_bg = 0.0;
-        double sumSq_bg = 0.0;
-
-        for (V3DLONG z = z_start; z < z_end; z++) {
-          for (V3DLONG y = y_start; y < y_end; y++) {
-            for (V3DLONG x = x_start; x < x_end; x++) {
-              V3DLONG idx = z * dim_X * dim_Y + y * dim_X + x;
-              if (originalData[idx] < threshold) {
-                sum_bg += originalData[idx];
-                sumSq_bg += originalData[idx] * originalData[idx];
-                backgroundCount++;
-              }
-            }
-          }
-        }
-
-        // If we found background voxels, calculate statistics
-        if (backgroundCount > 0) {
-          mean = sum_bg / backgroundCount;
-          double variance = (sumSq_bg / backgroundCount) - (mean * mean);
-          stdDev = sqrt(variance);
-
-          // Ensure minimum stdDev to avoid issues with uniform regions
-          if (stdDev < 1.0) stdDev = 1.0;
-        } else {
-          // If no background voxels found, use default values
-          mean = 10.0;
-          stdDev = 3.0;
-        }
-
-        // Store the mean and stdDev for this chunk
-        chunkStats[cz][cy][cx] = {mean, stdDev};
-      }
-    }
-  }
-
-  // Define blending parameters - use a smaller radius for more localized
-  // blending
-  double blendRadius = 1.0;
-
-  // Generate background values with blending between chunks
-  for (V3DLONG z = 0; z < dim_Z; z++) {
-    for (V3DLONG y = 0; y < dim_Y; y++) {
-      for (V3DLONG x = 0; x < dim_X; x++) {
-        // Get blended distribution parameters for this voxel
-        double blendedMean, blendedStdDev;
-        getBlendedDistributionParams(x, y, z, chunkStats, chunk_X, chunk_Y,
-                                     chunk_Z, num_chunks_X, num_chunks_Y,
-                                     num_chunks_Z, blendRadius, blendedMean,
-                                     blendedStdDev);
-
-        // Generate random value from the blended distribution
-        double u1 = std::rand() / (RAND_MAX + 1.0);
-        double u2 = std::rand() / (RAND_MAX + 1.0);
-
-        // Avoid log(0)
-        if (u1 < 1e-10) u1 = 1e-10;
-
-        double randStdNormal =
-            std::sqrt(-2.0 * std::log(u1)) * std::cos(2.0 * M_PI * u2);
-        double randNormal = blendedMean + blendedStdDev * randStdNormal;
-
-        // Clamp to valid unsigned char range [0,255]
-        int value = std::round(randNormal);
-        value = std::max(0, std::min(255, value));
-
-        backgroundArray[z * dim_X * dim_Y + y * dim_X + x] =
-            (unsigned char)value;
-      }
-    }
-  }
-
-  printf(
-      "Background generation complete with smooth blending (radius = %.1f).\n",
-      blendRadius);
-  return backgroundArray;
-}
-
-void free_mapped_arrays(unsigned char ***intensities,
-                        unsigned char ***segmentation, V3DLONG dim_Z,
-                        V3DLONG dim_Y) {
-  // Free intensities array
-  if (intensities) {
-    for (V3DLONG z = 0; z < dim_Z; z++) {
-      if (intensities[z]) {
-        for (V3DLONG y = 0; y < dim_Y; y++) {
-          if (intensities[z][y]) delete[] intensities[z][y];
-        }
-        delete[] intensities[z];
-      }
-    }
-    delete[] intensities;
-  }
-
-  // Free segmentation array
-  if (segmentation) {
-    for (V3DLONG z = 0; z < dim_Z; z++) {
-      if (segmentation[z]) {
-        for (V3DLONG y = 0; y < dim_Y; y++) {
-          if (segmentation[z][y]) delete[] segmentation[z][y];
-        }
-        delete[] segmentation[z];
-      }
-    }
-    delete[] segmentation;
-  }
-}
-
-void create_background(V3DPluginCallback2 &callback, QWidget *parent) {
-  // Old version that creates and displays an image directly for debugging
-
-  // Call the new version to get the background intensities
-  V3DLONG dimX, dimY, dimZ;
-  unsigned char *backgroundIntensities =
-      create_background(callback, parent, dimX, dimY, dimZ);
-
-  if (!backgroundIntensities) {
-    v3d_msg("Failed to generate background intensities.", parent);
-    return;
-  }
-
-  // Create new image for display
-  Image4DSimple *backgroundImage = new Image4DSimple();
-  backgroundImage->createBlankImage(dimX, dimY, dimZ, 1, V3D_UINT8);
-
-  // Get image properties from current image for consistency
-  v3dhandle curwin = callback.currentImageWindow();
-  if (curwin) {
-    Image4DSimple *p4DImage = callback.getImage(curwin);
-    if (p4DImage) {
-      backgroundImage->setOriginX(p4DImage->getOriginX());
-      backgroundImage->setOriginY(p4DImage->getOriginY());
-      backgroundImage->setOriginZ(p4DImage->getOriginZ());
-      backgroundImage->setRezX(p4DImage->getRezX());
-      backgroundImage->setRezY(p4DImage->getRezY());
-      backgroundImage->setRezZ(p4DImage->getRezZ());
-    }
-  }
-
-  // Copy the background intensities to the new image
-  unsigned char *backgroundData = backgroundImage->getRawData();
-  std::memcpy(backgroundData, backgroundIntensities, dimX * dimY * dimZ);
-
-  // Free the background intensities array
-  delete[] backgroundIntensities;
-
-  // Show the generated background
-  QString imageName = callback.getImageName(curwin);
-  v3dhandle newwin = callback.newImageWindow();
-  callback.setImage(newwin, backgroundImage);
-  callback.setImageName(newwin, imageName + "_background");
-  callback.updateImageWindow(newwin);
-
-  v3d_msg("Background image generated successfully.");
-}
-
-/**
- * @brief Calculate Otsu threshold for a histogram
- * @param hist Array of 256 histogram values
- * @param totalPixels Total number of pixels in the region
- * @return The calculated Otsu threshold value
- */
-int calculateOtsuThreshold(const int hist[256], int totalPixels) {
-  double sum = 0;
-  for (int t = 0; t < 256; t++) {
-    sum += t * hist[t];
-  }
-
-  double sumB = 0;
-  int wB = 0;
-  double varMax = 0;
-  int threshold = 0;
-
-  for (int t = 0; t < 256; t++) {
-    wB += hist[t];
-    if (wB == 0) continue;
-
-    int wF = totalPixels - wB;
-    if (wF == 0) break;
-
-    sumB += t * hist[t];
-    double mB = sumB / wB;
-    double mF = (sum - sumB) / wF;
-    double varBetween = (double)wB * wF * (mB - mF) * (mB - mF);
-
-    if (varBetween > varMax) {
-      varMax = varBetween;
-      threshold = t;
-    }
-  }
-
-  return threshold;
-}
-
-void getBlendedDistributionParams(
-    V3DLONG x, V3DLONG y, V3DLONG z,
-    const std::vector<std::vector<std::vector<std::pair<double, double>>>>
-        &chunkStats,
-    V3DLONG chunk_X, V3DLONG chunk_Y, V3DLONG chunk_Z, V3DLONG num_chunks_X,
-    V3DLONG num_chunks_Y, V3DLONG num_chunks_Z, double blendRadius,
-    double &blendedMean, double &blendedStdDev) {
-  // Calculate chunk coordinates and relative position
-  double cx_pos = (double)x / chunk_X;
-  V3DLONG cx = std::min(num_chunks_X - 1, (V3DLONG)cx_pos);
-
-  double cy_pos = (double)y / chunk_Y;
-  V3DLONG cy = std::min(num_chunks_Y - 1, (V3DLONG)cy_pos);
-
-  double cz_pos = (double)z / chunk_Z;
-  V3DLONG cz = std::min(num_chunks_Z - 1, (V3DLONG)cz_pos);
-
-  // Variables for weighted blending
-  double totalWeight = 0.0;
-  double weightedMean = 0.0;
-  double weightedVar = 0.0;
-
-  // Iterate over neighboring chunks within blend radius
-  for (int nz = std::max(0L, cz - (V3DLONG)blendRadius);
-       nz <= std::min(num_chunks_Z - 1, cz + (V3DLONG)blendRadius); nz++) {
-    for (int ny = std::max(0L, cy - (V3DLONG)blendRadius);
-         ny <= std::min(num_chunks_Y - 1, cy + (V3DLONG)blendRadius); ny++) {
-      for (int nx = std::max(0L, cx - (V3DLONG)blendRadius);
-           nx <= std::min(num_chunks_X - 1, cx + (V3DLONG)blendRadius); nx++) {
-        // Calculate distance to chunk center in chunk-space
-        double dx = (cx_pos - nx - 0.5);
-        double dy = (cy_pos - ny - 0.5);
-        double dz = (cz_pos - nz - 0.5);
-        double distSq = dx * dx + dy * dy + dz * dz;
-
-        // Skip chunks that are too far
-        if (distSq > blendRadius * blendRadius) continue;
-
-        // Calculate weight based on distance (quadratic falloff)
-        double weight = std::max(0.0, 1.0 - std::sqrt(distSq) / blendRadius);
-        weight = weight * weight;  // Square the weight for smoother falloff
-
-        // Get the statistics for this chunk
-        double mean = chunkStats[nz][ny][nx].first;
-        double stdDev = chunkStats[nz][ny][nx].second;
-
-        // Accumulate weighted statistics
-        weightedMean += weight * mean;
-        weightedVar += weight * stdDev * stdDev;  // Weighted variance
-        totalWeight += weight;
-      }
-    }
-  }
-
-  // Normalize by total weight
-  if (totalWeight > 0) {
-    weightedMean /= totalWeight;
-    weightedVar /= totalWeight;
-  } else {
-    // Fallback to central chunk if no weights (should not happen)
-    weightedMean = chunkStats[cz][cy][cx].first;
-    weightedVar = chunkStats[cz][cy][cx].second * chunkStats[cz][cy][cx].second;
-  }
-
-  // Set output parameters
-  blendedMean = weightedMean;
-  blendedStdDev = std::sqrt(weightedVar);
 }
