+++ released_plugins/v3d_plugins/MIND/soma_segmentation.pro
@@ -21,9 +21,13 @@ HEADERS += soma_segmentation_plugin.h
 HEADERS += ResolutionDialog.h
 HEADERS += cellSegmentation_plugin.h
 HEADERS += pcaAnalysis.h
+HEADERS += probability_model_vis.h
+HEADERS += soma_simulation.h
 SOURCES += soma_segmentation_plugin.cpp
 SOURCES += ResolutionDialog.cpp
 SOURCES += pcaAnalysis.cpp
+SOURCES += probability_model_vis.cpp
+SOURCES += soma_simulation.cpp
 SOURCES += $$VAA3DPATH/v3d_main/basic_c_fun/v3d_message.cpp
 SOURCES += $$VAA3DPATH/v3d_main/neuron_editing/v_neuronswc.cpp
 SOURCES += $$VAA3DPATH/v3d_main/basic_c_fun/basic_surf_objs.cpp
