+++ released_plugins/v3d_plugins/MIND/soma_segmentation_plugin.cpp
@@ -1,15 +1,12 @@
-/** soma_segmentation_plugin.cpp
- * This plugin supports the analaysis and segmentaiton of neuron somas in the
- * brain.
+/**
+ * This is the main file of the MIND plugin that supports the analaysis and
+ * segmentaiton of neuron somas in the brain. It contains the logic that
+ * controls which functions are called when the user selects a plugin option.
  *
- * It includes several functions:
- * - isotropic correction - to view 3D imagery of the brain isotropically
- * - soma segmentation - to segment individual somas using a user defined 3D
- * region-growing algorithm, and output a binary segmented image
- * - pc_analysis - to perform Principal Component Analysis on the input image
- * and output the results
+ * This file also contains the logic for isotropic correction, and intensity
+ * Weighted PCA on an image
  *
- * 2024-11-16: by ImagiNeuron: Shidan Javaheri, Siger Ma, Athmane Benarous and
+ * 2025-04-18: by ImagiNeuron: Shidan Javaheri, Siger Ma, Athmane Benarous and
  * Thibaut Baguette (McGill University)
  */
 
@@ -17,16 +14,15 @@
 
 #include <QApplication>
 #include <QFileDialog>
+#include <QFont>
 #include <QInputDialog>
 #include <QMessageBox>
 #include <QObject>
 #include <QPainter>
-#include <QFont>
 #include <algorithm>
 #include <cassert>
 #include <cmath>
 #include <cstring>
-#include <eigen/Dense>
 #include <fstream>
 #include <queue>
 #include <vector>
@@ -34,15 +30,16 @@
 #include "ResolutionDialog.h"
 #include "basic_4dimage.h"
 #include "basic_surf_objs.h"
-#include "v3d_message.h"
 #include "mainwindow.h"
+#include "v3d_message.h"
 
 using namespace std;
 
 /**
  * @brief Function to reconstruct somas
  *
- * Default function - copy format and add functionality afterwards
+ * Default function from plugin generator - copy format and add functionality
+ * afterwards
  *
  * @param callback - the V3D plugin callback interface
  * @param parent - the parent interface
@@ -131,7 +128,8 @@ MIND_4DImage *reconstruction_func(V3DPluginCallback2 &callback, QWidget *parent,
 QStringList SomaSegmentation::menulist() const {
   return QStringList() << tr("Isotropic Correction") << tr("Soma Segmentation")
                        << tr("PC Analysis") << tr("Visualize PCA")
-                       << tr("Visualize Probability Model") << tr("Probability Model Legend")
+                       << tr("Visualize Probability Model")
+                       << tr("Probability Model Legend")
                        << tr("Create Background") << tr("Simulate Somas")
                        << tr("about");
 }
@@ -542,725 +540,4 @@ void pca_func(V3DPluginCallback2 &callback, QWidget *parent, input_PARA &PARA,
   for (int i = 0; i < landmarkList.size(); i++) {
     analyzeSomaPCA(data1d, N, M, P, landmarkList[i], i + 1, savePath);
   }
-}
-
-std::tuple<unsigned char, unsigned char, unsigned char> colormap(double value) {
-  // Ensure value is between 0 and 1
-  value = std::clamp(value, 0.0, 1.0);
-  int index = static_cast<int>(value * (INFERNO_COLORMAP_SIZE - 1));
-
-  double r = std::clamp(std::get<0>(INFERNO_COLORMAP[index]), 0.0, 1.0);
-  double g = std::clamp(std::get<1>(INFERNO_COLORMAP[index]), 0.0, 1.0);
-  double b = std::clamp(std::get<2>(INFERNO_COLORMAP[index]), 0.0, 1.0);
-
-  return {static_cast<unsigned char>(r * 255), static_cast<unsigned char>(g * 255),
-          static_cast<unsigned char>(b * 255)};
-}
-
-void visualizeProbabilityModel_func(V3DPluginCallback2 &callback,
-                                    QWidget *parent) {
-  // Load data
-  QString filename = QFileDialog::getOpenFileName(
-      parent, "Open Probability Model", "", "Binary Files (*.bin)");
-
-  if (filename.isEmpty()) {
-    printf("No file selected.\n");
-    return;
-  }
-
-  std::vector<double> data;
-  V3DLONG dim_X, dim_Y, dim_Z;
-  cellSegmentation::class_segmentationMain::loadProbabilityModel(
-      filename.toStdString().c_str(), data, dim_X, dim_Y, dim_Z);
-
-  Image4DSimple *p4DImage = new Image4DSimple();
-  p4DImage->createBlankImage(dim_X, dim_Y, dim_Z, 3, V3D_UINT8);
-
-  double min = data[0];
-  double max = data[0];
-
-  for (V3DLONG i = 0; i < dim_X * dim_Y * dim_Z; i++) {
-    if (data[i] < min) min = data[i];
-    if (data[i] > max) max = data[i];
-  }
-
-  unsigned char *pixels = p4DImage->getRawData();
-  int channelSize = dim_X * dim_Y * dim_Z;
-  for (V3DLONG i = 0; i < dim_X * dim_Y * dim_Z; i++) {
-    auto [r, g, b] = colormap((data[i] - min) / (max - min));
-    pixels[i] = r;
-    pixels[i + channelSize] = g;
-    pixels[i + 2 * channelSize] = b;
-  }
-
-  v3dhandle newwin = callback.newImageWindow("Probability Model");
-  callback.setImage(newwin, p4DImage);
-
-  // open legend
-  probabilityModelLegend_func(callback, parent);
-}
-
-void probabilityModelLegend_func(V3DPluginCallback2 &callback, QWidget *parent) {
-  const int inner_width = 512; // Width of the color bar
-  const int inner_height = 60;  // Height of the color bar
-  const int padding_x = 40; // Padding around the color bar
-  const int padding_y = 15;
-  const int full_width = inner_width + 2 * padding_x;
-  const int full_height = inner_height + padding_y;
-
-  QImage legendImage(full_width, full_height, QImage::Format_ARGB32);
-  legendImage.fill(Qt::white);
-
-  // Draw the colormap in the top portion (say the top 20 pixels)
-  const int colorBarHeight = 30;
-  for (int x = padding_x; x < inner_width + padding_x; ++x) {
-      double value = double(x - padding_x) / (inner_width - 1); // [0,1]
-      auto [r, g, b] = colormap(value);
-      QColor color(static_cast<int>(r), static_cast<int>(g), static_cast<int>(b));
-  
-      // Fill a small vertical column of color
-      for (int y = 0; y < colorBarHeight; ++y) {
-          legendImage.setPixelColor(x, y, color);
-      }
-  }
-  
-  // Create a QPainter to draw numeric ticks/labels
-  QPainter painter(&legendImage);
-  painter.setPen(Qt::black);
-  painter.setFont(QFont("Arial", 12));
-  
-  // Define which ticks to draw. Here we do 0.0, 0.25, 0.5, 0.75, 1.0
-  QList<double> ticks = {0.0, 0.25, 0.5, 0.75, 1.0};
-  for (double t : ticks)
-  {
-      int xPos = int(t * (inner_width - 1) + padding_x);
-      // Vertical position below the color bar
-      int textY = colorBarHeight + 32; 
-      // Draw the numeric label
-      QString label = QString::number(t, 'f', 2); // e.g. "0.00", "0.25", ...
-      painter.drawText(xPos - 25, textY, label);  
-      // Optionally, draw a small tick line at each label
-      painter.drawLine(xPos, colorBarHeight, xPos, colorBarHeight + 10);
-  }
-  
-  // Set up a dialog to display our legend
-  QDialog *subWindow = new QDialog(callback.getVaa3DMainWindow());
-  subWindow->setWindowTitle("Colormap Legend");
-  subWindow->setWindowFlags(Qt::Window | Qt::WindowTitleHint | Qt::CustomizeWindowHint | Qt::WindowCloseButtonHint | Qt::MSWindowsFixedSizeDialogHint);
-
-  QLabel *label = new QLabel(subWindow);
-  label->setPixmap(QPixmap::fromImage(legendImage));
-  label->setAlignment(Qt::AlignCenter);
-  // If you do not want the label to stretch, omit "setScaledContents(true)"
-  label->setScaledContents(true);
-  
-  QVBoxLayout *layout = new QVBoxLayout(subWindow);
-  layout->addWidget(label);
-  subWindow->setLayout(layout);
-  
-  // Adjust as needed for an initial display size
-  subWindow->resize(600, 120);
-  subWindow->show();
-}
-
-/**
- * @brief overlay the ground truth data on the new simulated image. Dimesnions
- * calculated based on current image
- */
-void overlaySimulation(V3DPluginCallback2 &callback, QWidget *parent,
-                       unsigned char *binarySegImage,
-                       unsigned char *gradientImage,
-                       unsigned char *simulatedImage) {
-  v3dhandle curwin = callback.currentImageWindow();
-  v3dhandle newwin = callback.newImageWindow();
-
-  Image4DSimple *p4DImage = callback.getImage(curwin);
-  unsigned char *newData = new unsigned char[p4DImage->getTotalBytes() * 3];
-
-  memcpy(newData, simulatedImage, p4DImage->getTotalBytes());
-  memcpy(newData + p4DImage->getTotalBytes(), binarySegImage,
-         p4DImage->getTotalBytes());
-  memcpy(newData + p4DImage->getTotalBytes() * 2, gradientImage,
-         p4DImage->getTotalBytes());
-
-  Image4DSimple *newImage = new Image4DSimple;
-  newImage->setData(newData, p4DImage->getXDim(), p4DImage->getYDim(),
-                    p4DImage->getZDim(), p4DImage->getCDim() * 3, V3D_UINT8);
-
-  callback.setImage(newwin, newImage);
-}
-
-/**
- * @brief Function to simulate synthetic soma data
- *
- * @param callback - the V3D plugin callback interface
- * @param parent - the parent interface
- * @param PARA - the input parameters
- * @param bmenu - whether the function is being called from the menu
- */
-
-void simulate_soma_data(V3DPluginCallback2 &callback, QWidget *parent,
-                        input_PARA &PARA, bool bmenu) {
-  // Get current window and validate
-  v3dhandle curwin = callback.currentImageWindow();
-  if (!curwin) {
-    v3d_msg("No image window open!");
-    return;
-  }
-
-  Image4DSimple *p4DImage = callback.getImage(curwin);
-  if (!p4DImage) {
-    v3d_msg("Invalid image pointer!");
-    return;
-  }
-
-  // Get dimensions of current image
-  V3DLONG xDim = p4DImage->getXDim();
-  V3DLONG yDim = p4DImage->getYDim();
-  V3DLONG zDim = p4DImage->getZDim();
-
-  printf("\nStarting soma simulation...\n");
-  printf("Image dimensions: X=%ld, Y=%ld, Z=%ld\n", xDim, yDim, zDim);
-
-  // Get the current image name and path
-  QString imageName = callback.getImageName(curwin);
-  QString currentImagePath = QFileInfo(imageName).absolutePath();
-  QString baseImageName = QFileInfo(imageName).baseName();
-
-  /*
-   * Load original image data
-   */
-
-  unsigned char *originalData = p4DImage->getRawData();
-  int channel = 0;  // Default to first channel (index 0)
-
-  /*
-   * Load soma segmentation data
-   */
-
-  unsigned char *segData = nullptr;
-  V3DLONG sz[4];
-  int datatype = 0;
-  loadSegmentationFile(imageName, segData, sz, datatype, callback, parent);
-
-  /*
-   * Load segmentation image PCA and get distribution of soma properties
-   */
-
-  // Find the segmentation image PCA file
-  QString segPcaFileName =
-      modifyFilePathForTeraFly(imageName) + "_pca_binary_segmentation.csv";
-
-  // Check if the PCA file exists
-  if (!QFile::exists(segPcaFileName)) {
-    v3d_msg(
-        QString("segmentation image PCA file not found: %1\nPlease run soma "
-                "segmentation first.")
-            .arg(segPcaFileName));
-    delete[] segData;
-    return;
-  }
-
-  printf("\nLoading segmentation image PCA data from: %s\n",
-         segPcaFileName.toStdString().c_str());
-
-  // Load PCA data from CSV file
-  std::vector<double> eigenVectors;  // 9 eigenvector components
-  std::vector<double> centerCoords;  // CenterMassX, CenterMassY, CenterMassZ
-  std::vector<double> markerCoords;  // X, Y, Z
-  std::vector<double> somaRadii;     // Soma radii
-
-  std::ifstream segPcaFile(segPcaFileName.toStdString().c_str());
-  if (!segPcaFile.is_open()) {
-    v3d_msg("Could not open segmentation image PCA file!");
-    delete[] segData;
-    return;
-  }
-
-  std::string line;
-  // Skip header line
-  std::getline(segPcaFile, line);
-
-  // Read PCA data
-  int pcaRowCount = 0;
-  while (std::getline(segPcaFile, line)) {
-    std::stringstream ss(line);
-    std::string value;
-    std::vector<double> row;
-
-    while (std::getline(ss, value, ',')) {
-      row.push_back(std::stod(value));
-    }
-
-    // Expected columns in each CSV row:
-    //  0: SomaID
-    //  1: X
-    //  2: Y
-    //  3: Z
-    //  4: Radius
-    //  5: CenterMassX
-    //  6: CenterMassY
-    //  7: CenterMassZ
-    //  8: eigenvalue1
-    //  9: eigenvalue2
-    // 10: eigenvalue3
-    // 11: eigenvector1_x
-    // 12: eigenvector1_y
-    // 13: eigenvector1_z
-    // 14: eigenvector2_x
-    // 15: eigenvector2_y
-    // 16: eigenvector2_z
-    // 17: eigenvector3_x
-    // 18: eigenvector3_y
-    // 19: eigenvector3_z
-
-    markerCoords.push_back(row[1]);  // X
-    markerCoords.push_back(row[2]);  // Y
-    markerCoords.push_back(row[3]);  // Z
-
-    somaRadii.push_back(row[4]);  // Radius
-
-    centerCoords.push_back(row[5]);  // CenterMassX
-    centerCoords.push_back(row[6]);  // CenterMassY
-    centerCoords.push_back(row[7]);  // CenterMassZ
-
-    for (int iVec = 11; iVec < 20; iVec++) {
-      eigenVectors.push_back(row[iVec]);
-    }
-
-    pcaRowCount++;
-  }
-
-  printf("Loaded %d soma segmentation image PCA records\n", pcaRowCount);
-
-  // Calculate mean and standard deviation of center of mass coordinates,
-  // and eigenvectors
-  std::vector<double> meanCenter(3, 0.0);
-  std::vector<double> stdCenter(3, 0.0);
-  std::vector<double> meanEigenvectors(9, 0.0);
-  std::vector<double> stdEigenvectors(9, 0.0);
-
-  for (size_t i = 0; i < centerCoords.size(); i += 3) {
-    for (int j = 0; j < 3; j++) {
-      meanCenter[j] += centerCoords[i + j];
-    }
-  }
-
-  int numSomas = centerCoords.size() / 3;
-  for (int j = 0; j < 3; j++) {
-    meanCenter[j] /= numSomas;
-  }
-
-  for (size_t i = 0; i < centerCoords.size(); i += 3) {
-    for (int j = 0; j < 3; j++) {
-      stdCenter[j] += pow(centerCoords[i + j] - meanCenter[j], 2);
-    }
-  }
-
-  for (int j = 0; j < 3; j++) {
-    stdCenter[j] = sqrt(stdCenter[j] / numSomas);
-  }
-
-  for (size_t i = 0; i < eigenVectors.size(); i += 9) {
-    for (int j = 0; j < 9; j++) {
-      meanEigenvectors[j] += eigenVectors[i + j];
-    }
-  }
-
-  for (int j = 0; j < 9; j++) {
-    meanEigenvectors[j] /= numSomas;
-  }
-
-  // Calculate standard deviations
-  for (size_t i = 0; i < eigenVectors.size(); i += 9) {
-    for (int j = 0; j < 9; j++) {
-      stdEigenvectors[j] += pow(eigenVectors[i + j] - meanEigenvectors[j], 2);
-    }
-  }
-
-  for (int j = 0; j < 9; j++) {
-    stdEigenvectors[j] = sqrt(stdEigenvectors[j] / numSomas);
-  }
-
-  printf("\nPCA Statistics:\n");
-  printf("Mean center: (%.2f, %.2f, %.2f)\n", meanCenter[0], meanCenter[1],
-         meanCenter[2]);
-  printf("Std dev: (%.2f, %.2f, %.2f)\n", stdCenter[0], stdCenter[1],
-         stdCenter[2]);
-
-  /*
-   * Create synthetic somas and image
-   */
-
-  // Create output image
-  V3DLONG totalSize = xDim * yDim * zDim;
-  unsigned char *outSegData = new unsigned char[totalSize];
-  memset(outSegData, 0, totalSize);
-
-  // Create output image with original intensity values
-  unsigned char *outIntensityData =
-      create_background(callback, parent, xDim, yDim, zDim);
-
-  // Generate random positions and place synthetic somas
-  std::random_device rd;
-  std::mt19937 gen(rd());
-
-  // Ask user for the number of synthetic somas to generate
-  bool ok;
-  int numSynthetic =
-      QInputDialog::getInt(parent, "Synthetic Soma Generation",
-                           "Enter the number of synthetic somas to generate:",
-                           20,    // Default value
-                           1,     // Minimum value
-                           1000,  // Maximum value
-                           1,     // Step
-                           &ok);
-
-  if (!ok) {
-    // User canceled the dialog
-    printf("Synthetic soma generation canceled.");
-    delete[] segData;
-    return;
-  }
-
-  printf("\nGenerating %d synthetic somas...\n", numSynthetic);
-
-  int successfulPlacements = 0;
-
-  // Initialize simulatedLandmarks list
-  LandmarkList simulatedLandmarks;
-
-  // Helper function to check if two somas overlap
-  auto somasOverlap = [](const LocationSimple &s1,
-                         const LocationSimple &s2) -> bool {
-    // Calculate squared distance between centers
-    double dx = s1.x - s2.x;
-    double dy = s1.y - s2.y;
-    double dz = s1.z - s2.z;
-    double distSq = dx * dx + dy * dy + dz * dz;
-
-    // If distance is less than sum of radii, they overlap
-    double minDist = s1.radius + s2.radius;
-    return distSq < (minDist * minDist);
-  };
-
-  for (int i = 0; i < numSynthetic; i++) {
-    // Choose a random soma from the available ones for this synthetic soma
-    int randomSomaIndex = gen() % numSomas;
-
-    // Get the radius and calculate appropriate cube size
-    double radius = somaRadii[randomSomaIndex];
-    V3DLONG cubeSize = static_cast<V3DLONG>(
-        2.5 * radius);  // Use 2.5x radius to ensure we capture the whole soma
-
-    // Make sure cubeSize is odd for centering purposes
-    if (cubeSize % 2 == 0) cubeSize += 1;
-
-    // Set boundary margin based on the cube size
-    int boundaryMargin = cubeSize / 2;
-
-    // Generate random position using normal distribution
-    LocationSimple newSoma;
-    newSoma.radius = round(radius);
-    newSoma.name = qPrintable(QString("Sim_%1").arg(i + 1));
-    newSoma.comments = "";
-    newSoma.shape = pxSphere;
-
-    bool validPosition = false;
-    int maxAttempts = 100;
-    int attempts = 0;
-
-    while (!validPosition && attempts < maxAttempts) {
-      attempts++;
-      validPosition = true;
-
-      // Generate new potential position
-      for (int j = 0; j < 3; j++) {
-        std::normal_distribution<> d(meanCenter[j], stdCenter[j]);
-        double coord = round(d(gen));
-
-        // Store coordinate in new soma
-        if (j == 0) {
-          newSoma.x = coord;
-          // Check if position is within image boundaries
-          if (coord < boundaryMargin || coord > xDim - boundaryMargin) {
-            validPosition = false;
-            break;
-          }
-        } else if (j == 1) {
-          newSoma.y = coord;
-          // Check if position is within image boundaries
-          if (coord < boundaryMargin || coord > yDim - boundaryMargin) {
-            validPosition = false;
-            break;
-          }
-        } else if (j == 2) {
-          newSoma.z = coord;
-          // Check if position is within image boundaries
-          if (coord < boundaryMargin || coord > zDim - boundaryMargin) {
-            validPosition = false;
-            break;
-          }
-        }
-      }
-
-      // If position is within boundaries, check for overlap with existing somas
-      if (validPosition) {
-        // Check against all previously placed somas
-        for (int s = 0; s < simulatedLandmarks.size(); s++) {
-          if (somasOverlap(newSoma, simulatedLandmarks[s])) {
-            validPosition = false;
-            break;
-          }
-        }
-      }
-    }
-
-    if (!validPosition) {
-      printf(
-          "Failed to find valid non-overlapping position for soma %d after %d "
-          "attempts\n",
-          i + 1, maxAttempts);
-      continue;
-    }
-
-    // Add this soma to our landmarks list
-    simulatedLandmarks.append(newSoma);
-    successfulPlacements++;
-
-    // Generate random PCA values based on the distribution
-    double randomVec1[3], randomVec2[3], randomVec3[3];
-
-    Eigen::Matrix3d A;
-    for (int col = 0; col < 3; ++col) {
-      for (int row = 0; row < 3; ++row) {
-        std::normal_distribution<> dist(meanEigenvectors[col * 3 + row],
-                                        stdEigenvectors[col * 3 + row]);
-        A(row, col) = dist(gen);
-      }
-    }
-
-    // Perform QR decomposition to obtain orthogonal vectors
-    Eigen::HouseholderQR<Eigen::Matrix3d> qr(A);
-    Eigen::Matrix3d Q = qr.householderQ();
-
-    for (int i = 0; i < 3; ++i) {
-      randomVec1[i] = Q(i, 0);
-      randomVec2[i] = Q(i, 1);
-      randomVec3[i] = Q(i, 2);
-    }
-
-    // Extract a soma from segmentation data
-    // Get center of mass for the selected soma
-    V3DLONG sourceCenterX =
-        static_cast<V3DLONG>(centerCoords[randomSomaIndex * 3]);
-    V3DLONG sourceCenterY =
-        static_cast<V3DLONG>(centerCoords[randomSomaIndex * 3 + 1]);
-    V3DLONG sourceCenterZ =
-        static_cast<V3DLONG>(centerCoords[randomSomaIndex * 3 + 2]);
-
-    V3DLONG totalVoxels = cubeSize * cubeSize * cubeSize;
-    double *tempSegmentation = new double[totalVoxels];
-    double *tempIntensity = new double[totalVoxels];
-    memset(tempSegmentation, 0, totalVoxels * sizeof(double));
-    memset(tempIntensity, 0, totalVoxels * sizeof(double));
-
-    // Extract both segmentation and intensity data for the soma
-    for (int z = 0; z < cubeSize; z++) {
-      for (int y = 0; y < cubeSize; y++) {
-        for (int x = 0; x < cubeSize; x++) {
-          // Calculate positions relative to the soma's center
-          V3DLONG sourceX = sourceCenterX + x - cubeSize / 2;
-          V3DLONG sourceY = sourceCenterY + y - cubeSize / 2;
-          V3DLONG sourceZ = sourceCenterZ + z - cubeSize / 2;
-
-          // Target index in temporary buffer
-          int targetIdx = z * cubeSize * cubeSize + y * cubeSize + x;
-
-          // Check if coordinates are within the image bounds
-          if (sourceX >= 0 && sourceX < xDim && sourceY >= 0 &&
-              sourceY < yDim && sourceZ >= 0 && sourceZ < zDim) {
-            // Calculate index in the source images
-            V3DLONG sourceIdx =
-                sourceZ * xDim * yDim + sourceY * xDim + sourceX;
-
-            // Copy the segmentation value (0 or 255 for binary image)
-            tempSegmentation[targetIdx] = segData[sourceIdx] > 0 ? 1 : 0;
-
-            // Copy the original intensity value if this voxel is part of the
-            // soma
-            if (segData[sourceIdx] > 0) {
-              tempIntensity[targetIdx] =
-                  static_cast<double>(originalData[sourceIdx]);
-            }
-          }
-        }
-      }
-    }
-
-    // Apply random rotation to the synthetic soma
-    cellSegmentation::class_segmentationMain segMain;
-    segMain.rotateSegmentation(tempSegmentation, cubeSize, randomVec1,
-                               randomVec2, randomVec3);
-    segMain.rotateSegmentation(tempIntensity, cubeSize, randomVec1, randomVec2,
-                               randomVec3);
-
-    // Place rotated synthetic soma at generated position
-    int centerX = V3DLONG(newSoma.x);
-    int centerY = V3DLONG(newSoma.y);
-    int centerZ = V3DLONG(newSoma.z);
-
-    // Copy rotated soma to both output images
-    for (int z = 0; z < cubeSize; z++) {
-      for (int y = 0; y < cubeSize; y++) {
-        for (int x = 0; x < cubeSize; x++) {
-          int sourceIdx = z * cubeSize * cubeSize + y * cubeSize + x;
-
-          // Convert to output image coordinates
-          int targetX = centerX + x - cubeSize / 2;
-          int targetY = centerY + y - cubeSize / 2;
-          int targetZ = centerZ + z - cubeSize / 2;
-
-          if (targetX >= 0 && targetX < xDim && targetY >= 0 &&
-              targetY < yDim && targetZ >= 0 && targetZ < zDim) {
-            V3DLONG targetIdx =
-                targetZ * xDim * yDim + targetY * xDim + targetX;
-
-            // Only set voxel if rotated model indicates soma presence
-            if (tempSegmentation[sourceIdx] > 0) {
-              outSegData[targetIdx] = 255;  // Binary segmentation
-              outIntensityData[targetIdx] =
-                  static_cast<unsigned char>(std::round(std::min(
-                      255.0,
-                      std::max(
-                          0.0,
-                          tempIntensity[sourceIdx]))));  // Original intensity
-            }
-          }
-        }
-      }
-    }
-
-    delete[] tempSegmentation;
-    delete[] tempIntensity;
-  }
-
-  /*
-   * Save the synthetic soma segmentation images
-   */
-
-  // Generate timestamp for folder name
-  QDateTime currentTime = QDateTime::currentDateTime();
-  QString timestamp = currentTime.toString("yyyyMMdd_hhmmss");
-
-  // Create timestamped directory to store all output files
-  QString outputDirPath =
-      modifyFilePathForTeraFly(imageName) + "_simulation_" + timestamp;
-  QDir outputDir(outputDirPath);
-  if (!outputDir.exists()) {
-    outputDir.mkpath(".");
-    printf("Created output directory: %s\n",
-           outputDirPath.toStdString().c_str());
-  }
-
-  // Define output filenames in the new directory
-  QString outSegFileName = outputDirPath + "/simulated_segmentation.tif";
-  QString outIntensityFileName = outputDirPath + "/simulated_intensity.tif";
-
-  // Create dimension array for saving images
-  V3DLONG out_sz[4];
-  out_sz[0] = xDim;
-  out_sz[1] = yDim;
-  out_sz[2] = zDim;
-  out_sz[3] = 1;  // Single channel
-
-  // Save the binary segmentation image
-  simple_saveimage_wrapper(callback, outSegFileName.toStdString().c_str(),
-                           outSegData, out_sz, V3D_UINT8);
-
-  // Save the intensity image
-  simple_saveimage_wrapper(callback, outIntensityFileName.toStdString().c_str(),
-                           outIntensityData, out_sz, V3D_UINT8);
-
-  /*
-   * Save simulated soma landmarks as a marker file
-   */
-
-  QString markerFileName = outputDirPath + "/simulated_landmarks.marker";
-
-  FILE *fp = fopen(markerFileName.toStdString().c_str(), "w");
-  if (fp) {
-    // Write header
-    fprintf(fp, "#x, y, z, radius, shape, name, comment\n");
-
-    // Write each landmark
-    for (int i = 0; i < simulatedLandmarks.size(); i++) {
-      // Format: x,y,z,radius,shape,name,comment
-      fprintf(fp, "%ld,%ld,%ld,%ld,%ld,%s,%s\n",
-              V3DLONG(simulatedLandmarks.at(i).x),
-              V3DLONG(simulatedLandmarks.at(i).y),
-              V3DLONG(simulatedLandmarks.at(i).z),
-              V3DLONG(simulatedLandmarks.at(i).radius),
-              V3DLONG(simulatedLandmarks.at(i).shape),
-              simulatedLandmarks.at(i).name.c_str(),
-              simulatedLandmarks.at(i).comments.c_str());
-    }
-
-    fclose(fp);
-  } else {
-    printf("Error: Could not save marker file: %s\n",
-           markerFileName.toStdString().c_str());
-  }
-
-  /*
-   * Perform PCA analysis on the simulated somas
-   */
-
-  QString pcaSimulatedFileName =
-      outputDirPath + "/pca_simulated_segmentation.csv";
-
-  // Perform PCA analysis on each simulated soma
-  for (int i = 0; i < simulatedLandmarks.size(); i++) {
-    analyzeSomaPCA(outSegData, xDim, yDim, zDim, simulatedLandmarks[i], i + 1,
-                   pcaSimulatedFileName);
-  }
-
-  v3d_msg(QString("Simulation complete:\nGenerated %1/%2 synthetic "
-                  "somas. Files saved to directory: %3\n")
-              .arg(successfulPlacements)
-              .arg(numSynthetic)
-              .arg(outputDirPath));
-
-  delete[] segData;
-
-  /*
-   * Open new windows and display the synthetic soma data
-   */
-
-  unsigned char *gradientImage = new unsigned char[totalSize];
-  cellSegmentation cellSeg;
-  cellSeg.sobel3D(outSegData, gradientImage, xDim, yDim, zDim);
-
-  // overlay
-  overlaySimulation(callback, parent, outSegData, gradientImage,
-                    outIntensityData);
-  // // Create and show new window with binary simulated data. Now Obsolete
-  // Create an image for the binary and realistic simulation data
-  // Image4DSimple outSegImage;
-  // outSegImage.setData(outSegData, out_sz[0], out_sz[1], out_sz[2], out_sz[3],
-  //                     V3D_UINT8);
-  // v3dhandle segWin = callback.newImageWindow();
-  // callback.setImage(segWin, &outSegImage);
-  // callback.setImageName(segWin, outSegFileName);
-  // callback.updateImageWindow(segWin);
-
-  // // Create and show new window with intensity simulated data
-  // Image4DSimple outIntensityImage;
-  // outIntensityImage.setData(outIntensityData, out_sz[0], out_sz[1],
-  // out_sz[2],
-  //                           out_sz[3], V3D_UINT8);
-  // v3dhandle intensityWin = callback.newImageWindow();
-  // callback.setImage(intensityWin, &outIntensityImage);
-  // callback.setImageName(intensityWin, outIntensityFileName);
-  // callback.updateImageWindow(intensityWin);
 }
